# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management System API",
        "description": "API for managing hospital operations with role-based access control",
        "contact": {
            "email": "admin@hospital.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/doctor/report": {
            "get": {
                "tags": [
                    "👩‍⚕️ Doctor Services"
                ],
                "summary": "Create a new report for today",
                "description": "Creates a new daily report for the authenticated doctor",
                "operationId": "createDoctorReport",
                "responses": {
                    "201": {
                        "description": "Report created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success make report"
                                        },
                                        "data": {
                                            "properties": {
                                                "report": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "doctor_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "patient_count": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "income": {
                                                            "type": "number",
                                                            "example": 250
                                                        },
                                                        "report_date": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2025-04-30"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Doctor profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Report generation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "failed to make report"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Report for today already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/doctor/reports": {
            "get": {
                "tags": [
                    "👩‍⚕️ Doctor Services"
                ],
                "summary": "Get all doctor reports",
                "description": "Retrieves all reports generated by the authenticated doctor",
                "operationId": "getAllDoctorReports",
                "responses": {
                    "200": {
                        "description": "Reports retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success retrieve the report"
                                        },
                                        "data": {
                                            "properties": {
                                                "reports": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "doctor_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "patient_count": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "income": {
                                                                "type": "number",
                                                                "example": 250
                                                            },
                                                            "report_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2025-04-30"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Doctor profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve reports",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "failed to retrieve reports"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}